+++ 14th-team5-iOS/Data/Sources/Calendar/CalendarAPI/CalendarAPIs.swift
@@ -12,7 +12,6 @@ import Domain
 enum CalendarAPIs: API {
     case calendarInfo(String)
     case familySummaryInfo(String)
-    case familyCreatedAt(String)
     
     var spec: APISpec {
         switch self {
@@ -20,8 +19,6 @@ enum CalendarAPIs: API {
             return APISpec(method: .get, url: "\(BibbiAPI.hostApi)/calendar?type=MONTHLY&yearMonth=\(yearMonth)")
         case let .familySummaryInfo(familyId):
             return APISpec(method: .get, url: "\(BibbiAPI.hostApi)/families/\(familyId)/summary")
-        case let .familyCreatedAt(familyId):
-            return APISpec(method: .get, url: "\(BibbiAPI.hostApi)/families/\(familyId)/created-at")
         }
     }
 }
