+++ 14th-team5-iOS/Data/Sources/Calendar/Repository/CalendarRepository.swift
@@ -59,9 +59,4 @@ extension CalendarRepository {
         return calendarApiWorker.fetchFamilyStatisticsInfo(token: accessToken, familyId: familyId)
             .asObservable()
     }
-    
-    public func fetchFamilyCreatedAt() -> Observable<FamilyCreatedAtResponse?> {
-        return calendarApiWorker.fetchFamilyCreatedAt(token: accessToken, familyId: familyId)
-            .asObservable()
-    }
 }
