+++ 14th-team5-iOS/Data/Sources/Family/FamilyAPI/FamilyAPIWorker.swift
@@ -36,10 +36,6 @@ extension FamilyAPIs {
 }
 
 extension FamilyAPIWorker: SearchFamilyRepository {
-    public func getSavedFamilyMember(memberIds: [String]) -> [Domain.ProfileData]? {
-        return FamilyUserDefaults.loadMembersFromUserDefaults(memberIds: memberIds)
-    }
-    
     public func fetchFamilyMember(query: SearchFamilyQuery) -> Single<SearchFamilyPage?> {
         return Observable.just(())
             .withLatestFrom(self._headers)
@@ -129,5 +125,32 @@ extension FamilyAPIWorker: SearchFamilyRepository {
         let headers: [BibbiHeader] = [BibbiHeader.acceptJson, BibbiHeader.xAppKey, BibbiHeader.xAuthToken(accessToken)]
         
         return fetchFamilyMemberPage(spec: spec, headers: headers)
+    }
+    
+    private func fetchFamilyCreatedAt(spec: APISpec, headers: [BibbiHeader]) -> Single<FamilyCreatedAtResponse?> {
+        return request(spec: spec, headers: headers)
+            .subscribe(on: Self.queue)
+            .do {
+                if let str = String(data: $0.1, encoding: .utf8) {
+                    debugPrint("FamilyCreatedAt Fetch Result: \(str)")
+                }
+            }
+            .map(FamilyCreatedAtResponseDTO.self)
+            .catchAndReturn(nil)
+            .map { $0?.toDomain() }
+            .asSingle()
+    }
+    
+    public func fetchFamilyCreatedAt(token accessToken: String, familyId: String) -> Single<FamilyCreatedAtResponse?> {
+        let spec = FamilyAPIs.familyCreatedAt(familyId).spec
+        let headers: [BibbiHeader] = [.acceptJson, .xAppKey, .xAuthToken(accessToken)]
+        
+        return fetchFamilyCreatedAt(spec: spec, headers: headers)
+    }
+}
+
+extension FamilyAPIWorker {
+    public func getSavedFamilyMember(memberIds: [String]) -> [Domain.ProfileData]? {
+        return FamilyUserDefaults.loadMembersFromUserDefaults(memberIds: memberIds)
     }
 }
