+++ 14th-team5-iOS/Data/Sources/Family/FamilyAPI/FamilyAPIs.swift
@@ -13,6 +13,7 @@ public enum FamilyAPIs: API {
     case createFamily
     case invitationUrl(String)
     case familyMembers(FamilySearchRequestDTO)
+    case familyCreatedAt(String)
 
     var spec: APISpec {
         switch self {
@@ -22,6 +23,8 @@ public enum FamilyAPIs: API {
             return APISpec(method: .post, url: "\(BibbiAPI.hostApi)/links/family/\(familyId)")
         case let .familyMembers(query):
             return APISpec(method: .get, url: "\(BibbiAPI.hostApi)/members?type=FAMILY&page=\(query.page)&size=\(query.size)")
+        case let .familyCreatedAt(familyId):
+            return APISpec(method: .get, url: "\(BibbiAPI.hostApi)/families/\(familyId)/created-at")
         }
     }
 }
