+++ 14th-team5-iOS/Data/Sources/PostComment/PostCommentAPI/Repository/PostCommentRepositoryProtocol.swift
@@ -5,6 +5,7 @@
 //  Created by 김건우 on 1/17/24.
 //
 
+import Domain
 import Foundation
 
 import RxSwift
